requirements.txt
openai-whisper 
torch 
torchvision
torchaudio
pyannote.audio>=3.0
librosa
gradio
langchain
sentence_transformers 
PyMuPDF 
faiss-gpu-cu12
transformers 
accelerate 
safetensors
bitsandbytes
langchain_community
numpy
beautifulsoup4